device_orchestrator_Integration_testing/requirements.txt
pytest==8.3.4
pytest-asyncio==0.25.3
nats-py==2.9.0
python-dotenv==1.0.1

